fix(web): re-enable product registration page test
The test for the product registration alert was removed since it’s
already covered in its own test file.

As with other page tests, mocking the ProductRegistrationAlert component
was necessary to prevent the suite from rendering nothing, due to the
component getting stuck in the useIssues hook request. An alternative
would be to mock the useIssues hook itself, but it requires more mocking
for the same result. This is something to improve globally, but outside
the scope of this commit.