fix(web): prevent crash when deleting a drive
The interface was crashing when deleting a drive due to the
`RemoveDriveOption` component attempting to destructure an `undefined`
value after the delete action was performed.

Most likely, after the delete action, a parent component in the tree is
re-rendered, which causes `RemoveDriveOption` to re-render as well. As a
quick fix, this commit postpones the destructuring until the value
returned by the `useDrive` hook is known, returning early if it is
`undefined`. A more refined solution is welcome, but will be addressed
once the overall storage UI is more polished.