Enable some IceWM features in the config
Enable displaying the window title, the maximize/minimize buttons
and the window actions menu in the title bar.

YaST uses popup windows where these features are not wanted,
but Agama displays popups in the main browser window (so technically
they are not popups and we do not use any `alert()` or similar
functions). And these features might be useful when running xterm
debug window.