fix(web): add basic unit test for partition form
A quite generic test for at least ensuring it renders expected controls.
Some of these components deserves specific unit tests for testing its
specific behavior, but out of the scope of this commit.

To make things a bit better, some aria-labels have been added. But the
whole form still need quite some work for making it much more
accessible.

Last but not least, the commit also address an issue with the "Clear
mount point" button and add a test for it.