fix(web): use <a> instead of <button> for Page.Cancel
While writing a test for FirstUserForm, it was found that the "Cancel"
form button should be a link instead, as it navigates to another route
rather than performing an action.

After changing it to a link using the `component="a"` PatternFly prop,
it was discovered that React Testing Library couldn't find the "Cancel"
link using `getByRole`. According to
https://www.w3.org/TR/html-aria/#el-area-no-href, a link without an
`href` has a "generic" implicit role instead of the "link" role. This is
why `getByRole` didn't work: Page.Cancel was rendering a link without
the `href` attribute, but relying on the `onClick` event for navigation.

This commit fixes the Page.Cancel by relying on core/Link component for
producing a link instead of a button AND use `href` for navigation,
making the link more semantically correct and ensuring it can be
properly queried in tests.

See https://github.com/testing-library/react-testing-library/issues/1306