fix(web): expose password in root user form
After backend updates, defined root passwords are now exposed, enabling
more user-friendly forms. When editing root authentication methods with
a previously set plain password fields will be prefilled with their
values instead of being empty.

In addition to fixing a form issue when editing the root user with an
already set plain password, this improvement also helps users to check
previously entered root password insted of forcing to re-enter it in
case of doubt.

It also fixes the order of #then and #catch in the mutation promise,
similar to the change made in the FirstUser form at
48858e716780028dcac6821d0e9b31a6a9f0ed07.