feat(web): allow hostname to be set in registration form
The idea is to make it easier for users to set a hostname when
registering a product, as the product will be registered with the
specified hostname. Instead of adding warnings elsewhere, it’s more
helpful to let users set the hostname directly from the registration
form too, which will be prefilled with the current hostname value.

NOTE: This commit only updates the interface. The logic for setting the
hostname is not yet implemented. Tests are intentionally broken to
prevent overlooking it.