fix(web): add mising PF/FormHelperText wrapper
To be consistent with both, the rest of the code and PatternFly
examples.

Additionally, it wraps the inside text with a SubtleContent component to
provide a more subtle description. It can be done by using the
"indeterminate" variant in the HelperTextItem, but it adds an icon that
it does not fit in the current look & feel.