refactor(web): display current hostname and mode
Introduces a custom alert at the top of the hostname page to provide
users with both the current hostname value and the mode (permanent or
temporary). The goal is to ensure users not only know the current
hostname but also understand the meaning and implications of each mode,
helping them make an informed choice.

While there’s room for improvement, further refinements will be
addressed in future updates when moving toward a system page. This will
help unblock the current work and enable users to set the hostname via
the web interface as soon as possible.