refactor(web): use a checkbox for enabling encryption
Instead of using a switch. The previous implementation immediately
triggered the encryption settings when toggled, but this is no longer
the case. Switches should typically be used for immediate actions,
whereas encryption settings are now part of a form that must be
submitted before changes are applied.

Additionally, the way the form is rendered has been modified to show
fields only when necessary or requested, instead of just toggling their
enabled/disabled states. This ensures a more consistent interface by
displaying only required fields and aligns with the user
experience found on the authentication page.