fix(web): suggestion from code review
Use `model.Drive` instead of directly importing the `Drive` type. This
aims to make clear which Drive type is being used when reading, writing,
or maintaining the code.

However, StorageDevice is not under the `model` namespace (yet?) because
it doesn’t currently conflict with a type from `apiModel`. It would be
nice to find a better organization for these types, but there is neither
time nor a clear vision for how to structure this complex area at the
moment.