refactor(web): update behavior of "use disk" menu
No longer renders nothing when all available disks have been configured.
Instead, it now renders as disabled with an informative label.

Although previously discussed, this change is necessary to swap the
order of the actions menu in the config editor section. It feels more
natural to place it before the "More options" menu.

The approach of inverting the order to avoid altering the placement when
the "use disk" menu disappears seems to not scale well, especially in
the context of an empty config editor. To address this, this commit
adopts a similar approach to the partitions form page: disabling the
control and rendering an informative label. While this still leaves room
for improvements and fixes, particularly regarding a11y, it feels like a
step toward a more predictable and consistent interface.