refactor(web): swap order of config editor actions
Reorder the actions so "Use a disk" appears before "More options" as it
feels more natural.

As mentioned in a previous commit, the order was originally inverted to
keep the menus in the same place, because the "Use a disk" menu wasn't
rendered when all disks were used. However, this is no longer the case.