web: fix layout shift issues on progress screen
Simplified the layout to ensure that, from "md" screen sizes onwards,
the container fills the full width and the step occupies 33% of the
available height.

While the approach for displaying progress is not yet perfect, these
changes help minimize layout shifts and reduce the movement of elements
when steppers are re-rendered with different information.