fix(web): add new variant for Loading component
This commit introduces a new variant for the Loading component, allowing
it to be used without occupying the full screen or available space. This
should help reduce the sensation of loading messages jumping when
transitioning from a general loader to a more specific one, especially
when the latter is placed inside its container.

Additionally, the "full screen" variant has been re-centered,
effectively reverting commit 01604272b50e98500. However, this time, it
uses Flexbox with specific styling rather than relying on the Bullseye
component.