Fix History API loop (#2354)
## Problem

- When registering the product from CLI the web UI was broken in Firefox

![Image](https://github.com/user-attachments/assets/90c12b88-14ed-4c83-9ae7-9fa006b0084c)
- See https://github.com/agama-project/agama/issues/2274
- It was fixed by a workaround:
https://github.com/agama-project/agama/pull/2281
- But we should find the real fix and remove that ugly workaround

## Debugging

To debug the issue I have reverted the #2281 fix and added a small
snippet which tracks using the History API into the `web/src/index.tsx`
file:

```js
let counter = 1;
const original = window.history.pushState;
window.history.pushState = function (data, unused, url) {
  console.log("pushState: ", counter, data, url);
  counter += 1;
  original.apply(this, [data, unused, url]);
};
```

That revealed that the History API was switching quickly between the
product selection page and the probing progress page:


![agama-too-many-push-states-orig](https://github.com/user-attachments/assets/380ffc30-12e7-44fe-aff1-883feb329d0c)

The Firefox browser has a limit of 200 History API calls within 10
seconds. After that an error is reported and an exception is raised.
That exception causes a crash in the React router resulting in empty
screen.

The problematic code was in the main application router:

```tsx
if (selectedProduct === undefined && location.pathname !== PRODUCT.root) {
  return <Navigate to={PRODUCT.root} />;
}

if (phase === InstallationPhase.Config && isBusy && location.pathname !== PRODUCT.progress) {
  return <Navigate to={PRODUCT.progress} />;
}
```

The problem happened when the selected product was not updated yet
(`selectedProduct === undefined`) but probing was already in progress
(`isBusy === true`). In that case these conditions caused switching
between the two paths as fast as possible. On my machine it was
switching about 1000 times per second as displayed in the log above.

A similar problem was present in the product selection page.

## Solution

- Handle this case explicitly, if that happens display the loading page
until the selected product is updated.

## Testing

- Tested manually, registration from CLI works fine even without that
delay workaround
- There are just few calls to the History API

![agama-too-many-push-states-fix1](https://github.com/user-attachments/assets/4f161cc8-3554-4444-89a6-149c5371f1f9)