feat(web): disable l10n reuse until product is selected
The installer doesn't yet support persisting localization settings for
later application when a product is selected. Therefore, it's safer to
show an informative message that might help clarify or even state the
difference between interface and product localization, especially useful
when the installer options dialog is opened before a product has been
chosen.