fix(web): use :lang(xx) instead of :root[lang="xx"]
This change improves CSS language rule support by using the correct
pseudo-class for matching all elements with the `lang` attribute, rather
than limiting it to :root.

While the CSS :root pseudo-class refers to the <html> element, it has
been proven that :root[lang="zh"] is not fully working as expected,
maybe because the html `lang` attribute is being changed dynamically
after the page is loaded, preventing the associated styles from updating
correctly due to the precedence cascading order. Additionally, the
:lang("zh") directive targets not only elements with a language value of
"zh" but also more specific language variations such as "zh-Hant",
"zh-Hans", and "zh-TW", which is equivalent to using :root[lang|="zh"].

Therefore, it is preferable to use the :lang pseudo-class for better
compatibility and broader matching.

Further reads:

 * https://developer.mozilla.org/en-US/docs/Web/CSS/:lang
 * https://www.w3.org/International/questions/qa-css-lang
 * https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_cascade/Cascade#cascading_order
 * https://www.w3.org/TR/selectors/#root-pseudo
 * https://www.w3.org/TR/selectors/#lang-pseudo