feat(web): add Noto Sans as fallback font
Noto Sans is used for Cyrillic and CJK scripts. Previously, Verdana was
the only fallback font, which could lead to improperly rendered glyphs
when a Latin language was selected and Verdana was not available on the
system.

Since Agama serves all necessary fonts, it's safer to use Noto Sans as a
consistent fallback to ensure proper rendering across languages.