feat(web): make entire row in SelectableDataTable clickable
Allows users to select a row by clicking anywhere on it, not just the
radio input. Also adds a hover effect to help visually track data in an
attempt to improve readability for sight users.

While this change can improve convenience and ease of use for many, it
introduces non-native behavior since HTML table rows aren't inherently
meant to be interactive. This raises potential accessibility and
semantic concerns that need to be carefully reviewed before releasing it
to ensure the solution remains inclusive and robust across all use
cases.