live: fix_bootconfig.s390x: strip CDLABEL from the kiwi-generated grub config (bsc#1245453)
We are determining the Volume ID ourselves, and actually setting it
in /etc/cmdline.d during image creation phase.

Ensure we strip the root=live:CDLABEL cmdline entry added in the
kiwi-generated grub2 configuration.

During image build, having the entry in the cmdline is not an issue,
as the volid is passed directly to xorriso by the main fix_bootconfig
script.

However, when repacking the image with some tools such as mksusecd,
they might (correctly!) default to the specified label name in the
grub configuration, which in our case is not what we want.

Only s390x is affected as the other architectures's fix_bootconfig
scripts replace the whole grub configuration.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>