fix(web): use 'device' instead of 'interface'
As agreed, although these terms are somewhat interchangeable in the
network area, consistently using a single term improves usability.
'Device' was chosen over 'interface' because it’s more tangible and
closer to the physical world, making the UI easier to understand.

The term 'interface' is still used when listing device properties, to
help build a mental model for users and ease the connection between both
concepts.

Additionally, the label "Any interface" has been changed to "Unbound"
instead of "Any device", as it sounds simpler and more accurate.