Add the internal device id to the updated device (bsc#1247445). (#2680)
When a zFCPController is activated the UI is not refreshed properly
maintaining it as inactive.

- https://trello.com/c/5SeUZvuG
- https://bugzilla.suse.com/show_bug.cgi?id=1247445

The problem is that the changed device does not contain its id,
therefore we are not replacing it at all. For refreshing the **zFCP**
Controllers table, just notifying the device id and replacing the object
should be enough, but we need to retrieve also the disks associated with
the activated controller and this is done through the controllers query,
therefore, we invalidate it after any controllers change.

The only thing that could be improved is that it takes some time to get
the list of LUNs and the controllers selector updated.

- *Tested manually*