feat(web): detach expandable section in Storage
On the proposal page editor, the expandable section displaying the
partitions list has been detached to allow for custom toggle behavior.
Unfortunately, the ExpandableSection component doesn’t support the
following:

Accepting a function for toggleContent that would receive the isExpanded
value.

Accepting toggleContentExpanded or toggleContentCollapsed, as with the
toggleText prop.

As a result, the state had to be lifted to the parent component to
toggle between icons. Additionally, the default icon cannot be
customized or removed via props, so it had to be hidden using CSS.