fix(web): adjust three dots icon positioning
The adjustments made to the three dots icon have also been applied to
the vertical version. However, we decided to revert to the horizontal
icon as it is wider and may be more discoverable for users. We’ll
continue to experiment with this change and can switch back to the
vertical version if it proves to be more convenient based on user
feedback. Additionally, the horizontal version is less similar to the
icon commonly used for moving items, which isn't yet available in Agama
UI.