feat(web): start syncing UI state with URL
Begin managing some parts of the UI state (e.g. selected tabs and
expanded sections) via the URL, allowing to preserve it across
navigation and return users to the same view after actions like
submitting a form.

Initially, nuqs[1] was considered for query string handling, as it’s a
well-established library that simplifies the work a lot. However, it
does not officially support HashRouter in React Router 6.3[2]. While an
unofficial adapter exists[3], initial attempts to integrate it proved
too complex.

For now, the implementation uses useSearchParams and useLocation from
React Router. Though not as seamless as nuqs, this approach works for
the initial use case.

Next steps:

  * Reduce code duplication across components
  * Persist expanded/collapsed section states across navigation
  * Add a convenient way to automatically reset the query string when needed

[1] https://nuqs.dev/
[2] https://github.com/47ng/nuqs/issues/810
[3] https://github.com/offlegacy/nuqs-hash-router