Fixed creating `/LiveOS/.packages.json.gz` file in SLE builds (#2771)
## Problem

- The created `/LiveOS/.packages.json.gz` file in SLE builds is empty,
in openSUSE builds it contains correct content
- This is a fix up for #2717

## Details

- It turned out that for some reason the SLE build root does not contain
the jq package. The package is installed into the target Live ISO, but
it is missing in the build root. In openSUSE builds it is present,
either it uses some different defaults or some tool depend on that...

## Solution

- Run the `jq` tool from the installed chroot directory, we install it
in the Live ISO so it is guaranteed to be there.
- Alternatively it should be possible to tweak the OBS project config to
add more packages into the build root. But that would require to change
the project config in all places (OBS, IBS...). As we install the jq
tool into the image anyway then just use it from the image.
- Additionally fixed deleting repofiles, that fails if there is no
disabled repository


## Testing

- Tested manually, both SLE and openSUSE builds contain a correct
`.packages.json.gz` file