web: apply `{ suspense: true }` to all L10n queries
To prevent crashes when loading a route directly without prior
navigation from another route that has already loaded the data. This
approach may need to be reevaluated, as it might make sense to set
suspense: true by default for all queries in the future, instead of
passing it explicitly with each call.