web: add hook to track query refetch completion
TanStack Query does not provide a built-in way to notify when queries
have finished refetching after being invalidated. The library author
usually recommends relying on `useQuery` to react to query state
changes, which makes sense in most cases. However, in Agama there is a
use case that can be considered an exception to this rule, where
components should not directly depend on query state but only be
notified when their data is considered fresh after a certain moment.

This is the case for the core/Page component, which will be changed to
enter a "loading" state when receiving a signal from the Agama progress
API. After progress completion, a proposal update signal may be
received, causing related queries to be invalidated. At that point,
TanStack Query triggers a refetch, and the core/Page component must wait
until all affected queries have fully refetched before exiting the
loading state.

Exiting the loading state as soon as progress ends can result in
rendering stale data, which is then replaced moments later, causing
noticeable UI flicker. This hook allows tracking refetch completion so
the loading state is exited only when data is fresh.