web: fix core/Page tests
The tests were fixed by replacing `plainRender` with
`installerRender`where needed and reordering imports to ensure that
`test-utils` is imported before the component, allowing the centralized
react-routter mocks to work as expected.  Including `test-utils.tsx` in
`jest.config.js#setupFilesAfterEnv` was because it caused the
`ProductRegistrationAlertMock` to fail for unknown reasons.